pyproject.toml
[project]
name = "ar-488-esp32"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "circuit-synth",
    "websockets",
    "pillow",
    "mcp[cli]",
]

[project.scripts]
tektool = "host_software.tektool.cli:main"

[tool.uv.sources]
circuit-synth = { path = "circuit-synth", editable = true }

[tool.pytest.ini_options]
testpaths = ["host_software/tektool/tests"]
pythonpath = ["."]

[dependency-groups]
dev = [
    "pyserial>=3.5",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
